+++ Windows/Aoba.vcxproj
@@ -14,11 +14,11 @@
     <ClCompile Include="..\Aoba\src\AobaFactory\AobaFactory.cpp" />
     <ClCompile Include="..\Aoba\src\ColorTheme\ColorTheme.cpp" />
     <ClCompile Include="..\Aoba\src\CommandLine\CommandLine.cpp" />
-    <ClCompile Include="..\Aoba\src\AobaCore\ParallelTaskManager\ParallelTaskManager.cpp" />
-    <ClCompile Include="..\Aoba\src\AobaCore\AobaCore.cpp" />
-    <ClCompile Include="..\Aoba\src\AobaCore\AobaLifeCycle.cpp" />
-    <ClCompile Include="..\Aoba\src\AobaCore\PageManager\PageCycle.cpp" />
-    <ClCompile Include="..\Aoba\src\AobaCore\PageManager\PageManager.cpp" />
+    <ClCompile Include="..\Aoba\src\Core\ParallelTaskManager\ParallelTaskManager.cpp" />
+    <ClCompile Include="..\Aoba\src\Core\Core.cpp" />
+    <ClCompile Include="..\Aoba\src\Core\AobaLifeCycle.cpp" />
+    <ClCompile Include="..\Aoba\src\Core\PageManager\PageCycle.cpp" />
+    <ClCompile Include="..\Aoba\src\Core\PageManager\PageManager.cpp" />
     <ClCompile Include="..\Aoba\src\Imaging\Imaging.cpp" />
     <ClCompile Include="..\Aoba\src\Page\Page.cpp" />
     <ClCompile Include="..\Aoba\src\PixelUnit\PixelUnit.cpp" />
@@ -51,7 +51,7 @@
     <ClInclude Include="..\Aoba\include\Aoba\Constraint.hpp" />
     <ClInclude Include="..\Aoba\include\Aoba\DynamicColor.hpp" />
     <ClInclude Include="..\Aoba\include\Aoba\Aoba.hpp" />
-    <ClInclude Include="..\Aoba\include\Aoba\AobaCore.hpp" />
+    <ClInclude Include="..\Aoba\include\Aoba\Core.hpp" />
     <ClInclude Include="..\Aoba\include\Aoba\Imaging.hpp" />
     <ClInclude Include="..\Aoba\include\Aoba\InputEvent.hpp" />
     <ClInclude Include="..\Aoba\include\Aoba\KeyboardOperation.hpp" />
@@ -77,8 +77,8 @@
     <ClInclude Include="..\Aoba\include\Aoba\UIVStackView.hpp" />
     <ClInclude Include="..\Aoba\include\Aoba\UnifiedFont.hpp" />
     <ClInclude Include="..\Aoba\include\Aoba\WindowManager.hpp" />
-    <ClInclude Include="..\Aoba\src\AobaCore\ParallelTaskManager\ParallelTaskManager.hpp" />
-    <ClInclude Include="..\Aoba\src\AobaCore\PageManager\PageManager.hpp" />
+    <ClInclude Include="..\Aoba\src\Core\ParallelTaskManager\ParallelTaskManager.hpp" />
+    <ClInclude Include="..\Aoba\src\Core\PageManager\PageManager.hpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{7d81de8f-ea0c-4b54-9153-4c7424206620}</ProjectGuid>
